Add build + test + Itch.io deployment to CI (#1)
* Fix build warnings; add untested itch push

* Decompose script into separate jobs

* Add more build prereqs

* Run upload after test

* Use separate creds file; install wasm-opt

* Break wasm build into own step; debug butler path

* Use self-hosted runner, more debugging

* Remove prereq step, as we're using self-hosted runner now

* Remove even more prereqs

* Back to hosted runners

* Restore prereqs

* Try new PATH variable

* Refactor into itch installation then itch run steps

* More debug

* Even more debug

* Try using local butler rather than path

* Use separate cache process for wasm files

* even more debugging

* Debugging in script

* more debug logging

* Remove cd from upload script

* Only run itch uploads on pushes to main

+++ scripts/itch-upload.sh
@@ -8,21 +8,27 @@ WORKING_DIR="${1:-$(pwd)}"
 OUT_DIR="$WORKING_DIR/wasm_out"
 OUT_NAME="serverlooter"
 GIT_HASH=$(git rev-parse --short HEAD)
-OUT_ZIP=$OUT_DIR/$OUT_NAME-itch-$GIT_HASH.zip
+OUT_ZIP=$OUT_NAME-itch-$GIT_HASH.zip
+BUTLER_EXE="${2:-butler}"
 
 cp -r $WORKING_DIR/assets $OUT_DIR/assets
 
-cd $OUT_DIR
+echo "Verifying OUT_DIR before zip:"
+ls -l $OUT_DIR
 
-zip -r $OUT_ZIP \
-  serverlooter-${GIT_HASH}.js \
-  serverlooter_opt-${GIT_HASH}.wasm \
-  index.html \
-  assets
+zip -r $OUT_DIR/$OUT_ZIP \
+  $OUT_DIR/serverlooter-${GIT_HASH}.js \
+  $OUT_DIR/serverlooter_opt-${GIT_HASH}.wasm \
+  $OUT_DIR/index.html \
+  $OUT_DIR/assets
 
-cd $WORKING_DIR
+if [ $? -ne 0 ]; then
+  echo "failed to zip all requested files (errno=$?); aborting."
+  exit 1
+fi
+
 rm -r $OUT_DIR/assets
 
-butler push $OUT_ZIP $ITCHIO_USERNAME/$ITCHIO_GAME:$ITCHIO_CHANNEL
+$BUTLER_EXE push $OUT_DIR/$OUT_ZIP $ITCHIO_USERNAME/$ITCHIO_GAME:$ITCHIO_CHANNEL
 
-rm $OUT_ZIP
+rm $OUT_DIR/$OUT_ZIP
